move stream to its own submodule
