+++ pyacq/core/__init__.py
@@ -4,6 +4,5 @@ from .nodegroup import NodeGroup
 from .node import Node, WidgetNode
 from .nodelist import register_node_type
 from .manager import Manager, create_manager
-from .stream import OutputStream, InputStream
-from .sharedarray import SharedArray
+from .stream import OutputStream, InputStream, SharedArray
 from .tools import ThreadPollInput, ThreadPollOutput, StreamConverter, StreamSplitter
