+++ pyacq/core/stream/sharedarray.py
